+++ murrelet_livecode_derive/src/lib.rs
@@ -11,7 +11,6 @@ mod derive_lazy;
 mod derive_livecode;
 mod derive_nestedit;
 mod parser;
-mod toplevel;
 
 use darling::FromDeriveInput;
 use derive_graphics_trait::impl_graphics_trait;
@@ -22,7 +21,6 @@ use parser::{GenFinal, LivecodeReceiver};
 use proc_macro::TokenStream;
 use proc_macro2::TokenStream as TokenStream2;
 use syn::parse_macro_input;
-use toplevel::{impl_all_the_traits, top_level_livecode, top_level_livecode_json};
 
 use quote::quote;
 
@@ -89,25 +87,6 @@ pub fn murrelet_livecode_derive_nestedit(input: TokenStream) -> TokenStream {
     let ast = parse_macro_input!(input as syn::DeriveInput);
     let ast_receiver = LivecodeReceiver::from_derive_input(&ast).unwrap();
     nestedit_parse_ast(ast_receiver.clone()).into()
-}
-
-// todo, this is if we need to load config
-#[proc_macro_derive(TopLevelLiveCode, attributes(livecode))]
-pub fn murrelet_livecode_top_level_livecode(input: TokenStream) -> TokenStream {
-    let ast = parse_macro_input!(input as syn::DeriveInput);
-    top_level_livecode(ast.ident).into()
-}
-
-#[proc_macro_derive(TopLevelLiveCodeJson, attributes(livecode))]
-pub fn murrelet_livecode_top_level_livecode_json(input: TokenStream) -> TokenStream {
-    let ast = parse_macro_input!(input as syn::DeriveInput);
-    top_level_livecode_json(ast.ident).into()
-}
-
-#[proc_macro_derive(LiveCoderTrait, attributes(livecode))]
-pub fn murrelet_livecode_livecoder_traits(input: TokenStream) -> TokenStream {
-    let ast = parse_macro_input!(input as syn::DeriveInput);
-    impl_all_the_traits(ast.ident).into()
 }
 
 #[proc_macro_derive(LiveGraphics, attributes(graphics))]
